+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @quarkcss/react
 
+## 0.5.1
+
+### Patch Changes
+
+- Fix multiline when base string instead of config
+- Updated dependencies
+  - @quarkcss/core@0.5.1
+
 ## 0.5.0
 
 ### Minor Changes